requirements.txt
ultralytics
flask
pillow